Make Subject with multicasting and fuse() method
Construct Subject with subject transmitter SubjectTx and
subject receiver SubjectRx with multicasting capability.
Add fuse() method to Observable and to SubjectRx objects which discards
all futher emitting after complete() is called.
Add defuse() method to Observable and to SubjectRx objects so that
emitting continues even if complete() is called, that is the
default behaviour.
Change Subcriber to stop emitting if error() is called.

Also change signature of Subscriber error() to take error type wrapped
in std::rc::Rc so references to error can be cloned and passed to each
registered Subcriber in Subject when Subject error() is called.