Enhance `take()`, add `merge()` and `merge_one()`
The previous version of the `take()` operator spawned an OS thread to
await a signal for unsubscribing when observable emits. The enhanced
`take()` operator now determines whether to use Tokio tasks or OS
threads for awaiting unsubscribing signals, based on the observable's
execution context.

Also, include optimizations of the `take()` operator for subjects in
this version. When `take()` is used directly on a subject variant,
unnecessary spawning of Tokio tasks or OS threads for awaiting
unsubscribing signals is avoided, resulting in more efficient resource
usage.

However, if `take()` is applied directly to any subject variant, it
does not unsubscribes and release resources held by the subject,
allowing for repeated subscriptions. If another operator is used before
`take()`, the first operator wraps the subject into an observable. In
that case `take()` will be called on observable wrapping the subject
and it will unsubscribe releasing the resources held by that subject.

Additionally, in this update introduce new methods, `merge()` and
`merge_one()`. These methods are used for turning two or more
observables into a single observable.

To enhance usability, implement the `From` trait to allow seamless
conversion of subject receivers to observables.

Remove `fuse()` and `defuse()` methods from subject variants as they
are redundant. Subjects automatically clear their registered
subscribers upon completion, making these methods unnecessary.