Implement `connectable()` operator
Add the `connectable()` operator in the `ObservableExt` trait to create
a `Connectable` observable.

Introduce the new `Connectable` observable type with a `connect()`
method to establish a connection and multicast emissions to multiple
observers.

This feature enhances the flexibility of observables by allowing them
to defer emissions until explicitly connected, enabling multiple
subscribers to receive the same emissions.