Implement `first()` operator
Add the `first()` operator to select the first item in an observable
sequence based on the provided `predicate`. If no item satisfies the
`predicate`, the operator emits a default value, if provided.
Otherwise, it emits an `EmptyError`.

The `first()` operator unsubscribes from background emissions once it
takes the first item that satisfies the `predicate`.